Change request method strategy when checking the access token
There is a known issue with the LinkedIn API (and other oAuth2 APIs too)
when sending an access token provided in the body of a POST request
that returns random REVOKED_ACCESS_TOKEN responses.

The workaround is to provide the same access token via query params of a
POST request instead. Thankfully, the `oauth2` gem already as a setting for
that matter, and it's implemented in this commit.